% Manual entry for compor_matriculas_etapa_projetadas
\name{compor_matriculas_etapa_projetadas}
\alias{compor_matriculas_etapa_projetadas}
\title{Compor matriculas projetadas por etapa de ensino}
\usage{
compor_matriculas_etapa_projetadas(indicadores_matricula_projetados)
}
\arguments{
\item{indicadores_matricula_projetados}{Data frame de indicadores e
matriculas projetadas por faixa etaria, como retornado por
\code{\link[=projetar_indicadores_matricula]{projetar_indicadores_matricula()}}.}
}
\value{
Data frame com matriculas projetadas por UF, etapa de ensino e ano,
com o total de matriculas em cada etapa.
}
\description{
Compor matriculas projetadas por etapa de ensino.
}
