% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionarios.R
\docType{data}
\name{dicionario_indicadores_matricula_observados}
\alias{dicionario_indicadores_matricula_observados}
\title{Dicionario dos indicadores observados de matricula}
\format{
Um data frame com 23 linhas e 6 colunas:
\describe{
\item{variavel}{Nome da variavel.}
\item{descricao}{Descricao curta da variavel.}
\item{unidade}{Unidade ou tipo de valor.}
\item{fonte_dados}{Fonte geral dos dados ou indicacao de calculo pelo pacote.}
\item{tabela_origem}{Tabela, conjunto de tabelas ou regra de derivacao de origem.}
\item{observacao}{Observacao sobre calculo, fonte ou aplicabilidade.}
}
}
\usage{
dicionario_indicadores_matricula_observados
}
\description{
Dicionario das variaveis retornadas por
\code{\link[=calcular_indicadores_matricula_observados]{calcular_indicadores_matricula_observados()}}.
}
\keyword{datasets}
