% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{ler_populacao_municipio_2025}
\alias{ler_populacao_municipio_2025}
\title{Ler estimativa populacional municipal de 2025}
\usage{
ler_populacao_municipio_2025()
}
\value{
Data frame com populacao estimada por UF, municipio e ano.
}
\description{
Ler estimativa populacional municipal de 2025
}
