% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{ler_projecao_populacional_ibge}
\alias{ler_projecao_populacional_ibge}
\title{Ler dados de projecao populacional do IBGE}
\usage{
ler_projecao_populacional_ibge(con = NULL)
}
\arguments{
\item{con}{Conexao opcional com MotherDuck. Se omitida, uma nova conexao sera
aberta e fechada automaticamente.}
}
\value{
Data frame com populacao projetada por UF, etapa de ensino e ano.
}
\description{
Ler dados de projecao populacional do IBGE
}
