% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_matricula.R
\name{percentual_matriculas_fora_faixa}
\alias{percentual_matriculas_fora_faixa}
\title{Calcular percentual de matriculas fora da faixa etaria}
\usage{
percentual_matriculas_fora_faixa(matriculas_fora_faixa, total_matriculas)
}
\arguments{
\item{matriculas_fora_faixa}{Numero de matriculas fora da faixa etaria
adequada.}

\item{total_matriculas}{Total de matriculas.}
}
\value{
Vetor numerico com o percentual de matriculas fora da faixa etaria.
}
\description{
Calcular percentual de matriculas fora da faixa etaria
}
