% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_matricula.R
\name{taxa_bruta_matricula}
\alias{taxa_bruta_matricula}
\title{Calcular taxa bruta de matricula}
\usage{
taxa_bruta_matricula(matriculas_faixa_etaria, populacao_faixa_etaria)
}
\arguments{
\item{matriculas_faixa_etaria}{Numero de matriculas na faixa etaria.}

\item{populacao_faixa_etaria}{Populacao na faixa etaria.}
}
\value{
Vetor numerico com a taxa bruta de matricula em percentual.
}
\description{
Calcular taxa bruta de matricula
}
