% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_matricula.R
\name{taxa_liquida_matricula}
\alias{taxa_liquida_matricula}
\title{Calcular taxa liquida de matricula}
\usage{
taxa_liquida_matricula(matriculas_faixa_adequada, populacao_faixa_adequada)
}
\arguments{
\item{matriculas_faixa_adequada}{Numero de matriculas na faixa etaria
adequada.}

\item{populacao_faixa_adequada}{Populacao na faixa etaria adequada.}
}
\value{
Vetor numerico com a taxa liquida de matricula em percentual.
}
\description{
Calcular taxa liquida de matricula
}
