% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{ler_matricula_faixaetaria}
\alias{ler_matricula_faixaetaria}
\title{Ler dados de matricula por faixa etaria}
\usage{
ler_matricula_faixaetaria(con = NULL)
}
\arguments{
\item{con}{Conexao opcional com MotherDuck. Se omitida, uma nova conexao sera
aberta e fechada automaticamente.}
}
\value{
Data frame com matriculas por UF, etapa de ensino, ano e faixa etaria.
}
\description{
Ler dados de matricula por faixa etaria
}
